# docs/swagger.yaml
definitions:
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handlers.SuccessResponse:
    properties:
      data:
        description: Опциональные данные
      message:
        type: string
    type: object
...
  models.Schedule:
    properties:
      created_at:
        type: string
      day_of_week:
        type: integer
      end_time:
        type: string
      group_id:
        type: integer
      group_name:
        type: string
      id:
        type: integer
      location:
        type: string
      start_time:
        type: string
      subject_id:
        type: integer
      subject_name:
        type: string
      teacher_id:
        type: integer
      teacher_name:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/schedule:
    get:
      consumes:
      - application/json
      description: Возвращает список всех занятий
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            items:
              $ref: '#/definitions/models.Schedule'
            type: array
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Получить общее расписание
      tags:
      - Schedules
    post:
      consumes:
      - application/json
      description: Создаёт новое занятие в расписании
      parameters:
      - description: Данные о занятии
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/models.Schedule'
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Создать занятие
      tags:
      - Schedules
  /api/schedule/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет занятие по его ID
      parameters:
      - description: ID занятия
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Удалить занятие
      tags:
      - Schedules
    get:
      consumes:
      - application/json
      description: Возвращает расписание для конкретной группы или преподавателя
      parameters:
      - description: ID группы или преподавателя
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            items:
              $ref: '#/definitions/models.Schedule'
            type: array
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Расписание не найдено
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Получить расписание по ID группы/преподавателя
      tags:
      - Schedules
    put:
      consumes:
      - application/json
      description: Обновляет информацию о существующем занятии
      parameters:
      - description: ID занятия
        example: 1
        in: path
        name: id
        required: true
        type: integer
      - description: Обновлённые данные о занятии
        in: body
        name: updates
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Обновить занятие
      tags:
      - Schedules
...
swagger: "2.0"